+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - `models.TextChoices` refactored for better code readability
+- Models refactored
 
 
 ## [1.2.0] - 2022-01-12